Speed up complex.pow when both base and exponent are real; when only the exponent is real; and when the base is Euler's number
These are some pretty common cases which appear in many formulas. The speed ups are pretty significant. According to my measurements (using the timeit library) when both base and exponent are real the speedup is ~2x; when only the exponent is real it is ~1.5x and when the base is Euler's number it is ~2x.

There is no measurable difference when using other exponents which makes sense since I refactored the code a little to reduce the total number of branches that are needed to get to the final "fallback" branch, and those branches have less comparisons (anecdotally the fallback case feels slightly faster, but the improvement is quite small to be properly measured, perhaps in the order of 3 or 4%).