Don't crash nimsuggest in case of internal compiler error. Raise a recoverable error instead.
Even though internal compiler errors should not happen and should be fixed,
nimsuggest doesn't exactly reproduce the compiler behaviour, because it ignores
fatal errors and continues compilation. This causes extra internal compiler
errors, which wouldn't happen during normal compilation. It is not good when
these internal errors cause nimsuggest to crash, since it's expected to work on
unfinished and/or broken code inside an IDE and it should have good error
recovery and still provide hints and code tools support for the parts that it
has successfully compiled, and this data should be reliable up until the first
error, at least.