Fixes an issue where two types can collide in the cgen
When generating a header with `--header` and then using it from another Nim project to import the generated types, generic instances can collide and produce an error i.e. `typedef MyGeneric[TypeA] typename` `typedef MyGeneric[TypeB] typename` `signature` here cant be the same as they are different types.  This issue can be avoided by using the Nim type as part of the name which is what PR does.